+++ packages/tail/src/System/Tail.hs
@@ -1,5 +1,6 @@
 {-# LANGUAGE OverloadedRecordDot #-}
 {-# LANGUAGE RecordWildCards #-}
+{-# LANGUAGE TemplateHaskell #-}
 
 {- |
 File tailing library with multi-subscriber support.
@@ -37,6 +38,7 @@ import Control.Concurrent.STM.CircularBuffer qualified as CB
 import System.Directory (doesFileExist)
 import System.IO (hGetLine)
 import System.Process (CreateProcess (..), ProcessHandle, StdStream (..), createProcess, proc, terminateProcess, waitForProcess)
+import System.Which (staticWhich)
 
 -- | Represent `tail -f`'ing a file in Haskell
 data Tail = Tail
@@ -58,8 +60,16 @@ data Tail = Tail
 The tail process starts immediately and begins reading from the file.
 New subscribers will receive a ring buffer containing the last @bufferSize@ lines.
 -}
-tailFile :: (HasCallStack) => FilePath -> Int -> FilePath -> IO Tail
-tailFile tailBin bufferSize filePath = do
+
+{- | Path to the `tail` executable
+
+This should be available in the PATH, thanks to Nix and `which` library.
+-}
+tailBin :: FilePath
+tailBin = $(staticWhich "tail")
+
+tailFile :: (HasCallStack) => Int -> FilePath -> IO Tail
+tailFile bufferSize filePath = do
   unlessM (doesFileExist filePath) $ error $ "File does not exist: " <> toText filePath
   queues <- newTVarIO mempty
   stop <- newEmptyTMVarIO
@@ -67,7 +77,7 @@ tailFile tailBin bufferSize filePath = do
   ringBuffer <- atomically $ CB.new bufferSize
   let t = Tail {..}
   -- Start the tail process immediately
-  void $ async $ tailRun tailBin t
+  void $ async $ tailRun t
   pure t
 
 {- | Signal the tail process to stop reading the file.
@@ -78,8 +88,8 @@ tailStop :: Tail -> IO ()
 tailStop t = do
   atomically $ putTMVar t.stop ()
 
-tailRun :: FilePath -> Tail -> IO ()
-tailRun tailBin t = do
+tailRun :: Tail -> IO ()
+tailRun t = do
   -- Start the tail -F process (show entire file from beginning)
   let createProc = (proc tailBin ["-F", "-n", "+1", t.filePath]) {std_out = CreatePipe}
   (_, Just hout, _, ph) <- createProcess createProc